fix scaler datatype

+++ main/include/eudaq/FileWriterTreeDRS4.hh
@@ -200,7 +200,7 @@ namespace eudaq {
         std::vector<uint16_t> wf_thr;
 
         //tu
-        std::vector<uint16_t> * v_scaler;
+        std::vector<uint32_t> * v_scaler;
 
     };
 }
